Renamed the file and modified according to the GUI designing
